still not working like I'd like it to (now it drops very crammed timelines)
I need to figure out how to modify the renderer to look at the list of dateSeq and if some # of them are NA (maybe 75%), change the label type to %m/%y which would be two digit month and year. feels like whenever I muck with that I can't solve the errors it throws. Will keep poking at it tomorrow.