adding in another approval type not entirely sure how to include properly in the index part of utils-json, nor the extendToWholeDays
I think these two spots assume Working is God, instead of Admin Read-Only. If that's true, I can make a better guess, but should confirm first.

This is just kind of a fun add-on, since Laura accidentally included this type in the ticket initially, but also gives us a good test to ask, how do we add another approval type, when the time comes?